Improved Encapsulation
